+++ src/applications/mhv/secure-messaging/containers/FolderThreadListView.jsx
@@ -2,6 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { useLocation, useParams } from 'react-router-dom';
 import PropTypes from 'prop-types';
+import { VaPagination } from '@department-of-veterans-affairs/component-library/dist/react-bindings';
 import {
   DefaultFolders as Folders,
   Alerts,
@@ -14,6 +15,7 @@ import AlertBackgroundBox from '../components/shared/AlertBackgroundBox';
 import { closeAlert } from '../actions/alerts';
 import ThreadsList from '../components/ThreadList/ThreadsList';
 import { getListOfThreads, clearListOfThreads } from '../actions/threads';
+import ThreadListSort from '../components/ThreadList/ThreadListSort';
 
 const FolderThreadListView = props => {
   const { testing } = props;
@@ -21,7 +23,7 @@ const FolderThreadListView = props => {
   const [folderId, setFolderId] = useState(null);
   const error = null;
   // Currently the pagination has a bug that doesnt return the correct amount of results per page, as a temporary solution, the threadsPerPage is set to 100 to cover most folders without needing pagination.
-  const threadsPerPage = 100;
+  const threadsPerPage = 10;
   const threads = useSelector(state => state.sm.threads?.threadList);
   const folder = useSelector(state => state.sm.folders.folder);
   const location = useLocation();
@@ -30,33 +32,44 @@ const FolderThreadListView = props => {
   const [sortOrder, setSortOrder] = useState(threadSortingOptions.DESCENDING);
   const [sortBy, setSortBy] = useState(threadSortingOptions.SORT_BY_SENT_DATE);
 
-  const handleThreadApiCall = () => {
+  const MAX_PAGE_LIST_LENGTH = 5;
+
+  const handleSortCallback = () => {
+    setPageNum(1);
     dispatch(
-      getListOfThreads(folderId, threadsPerPage, pageNum, sortBy, sortOrder),
+      getListOfThreads(folderId, threadsPerPage, 1, sortBy, sortOrder),
       true,
+    );
+  };
+
+  const handlePagination = page => {
+    setPageNum(page);
+    dispatch(
+      getListOfThreads(folderId, threadsPerPage, page, sortBy, sortOrder),
     );
   };
 
   useEffect(
     () => {
       if (folderId !== null) {
-        dispatch(retrieveFolder(folderId)).then(() => {
-          dispatch(
-            getListOfThreads(
-              folderId,
-              threadsPerPage,
-              pageNum,
-              sortBy,
-              sortOrder,
-            ),
-          );
-        });
+        setPageNum(1);
+        dispatch(retrieveFolder(folderId));
+        dispatch(
+          getListOfThreads(
+            folderId,
+            threadsPerPage,
+            1, // pageNum
+            sortBy,
+            sortOrder,
+            true,
+          ),
+        );
       }
       // on component unmount, clear out threads reducer to prevent from
       // previous threads results flashing when navigating between messages
       return () => dispatch(clearListOfThreads());
     },
-    [folderId, dispatch, pageNum],
+    [folderId, dispatch],
   );
 
   useEffect(
@@ -163,7 +176,6 @@ const FolderThreadListView = props => {
           pageNum={pageNum}
           setSortOrder={setSortOrder}
           setSortBy={setSortBy}
-          handleThreadApiCall={handleThreadApiCall}
           threadsPerPage={threadsPerPage}
         />
       );
@@ -179,7 +191,22 @@ const FolderThreadListView = props => {
         {folder?.folderId !== undefined && (
           <>
             <FolderHeader folder={folder} />
+            <ThreadListSort
+              defaultSortOrder={threadSortingOptions.DESCENDING}
+              setSortOrder={setSortOrder}
+              setSortBy={setSortBy}
+              sortCallback={handleSortCallback}
+            />
             {content()}
+            {threads?.length > 1 && (
+              <VaPagination
+                onPageSelect={e => handlePagination(e.detail.page)}
+                page={pageNum}
+                pages={Math.ceil(threads[0]?.threadPageSize / threadsPerPage)}
+                maxPageListLength={MAX_PAGE_LIST_LENGTH}
+                showLastPage
+              />
+            )}
           </>
         )}
       </div>
