+++ src/applications/mhv/secure-messaging/sass/folder-thread-list-view.scss
@@ -1,40 +1,16 @@
 @import "~@department-of-veterans-affairs/formation/sass/shared-variables";
 
 .message-count {
-    font-weight: normal;
+  font-weight: normal;
 }
 .thread-list-draft {
-    color: $color-secondary-darkest;
+  color: $color-secondary-darkest;
 }
 .thread-list {
-    width: 100%;
-    // @media (max-width: $small-screen) {
-        word-wrap: break-word;
-    //   }
-  .thread-list-sort {
-    width: 100%;
-
-    display: flex;
-    align-items: flex-end;
-
-    va-select {
-      margin-right: 16px;
-      min-width: 220px;
-    }
-
-    va-button {
-      display: contents;
-      height: 37px;
-      padding: 0;
-    }
-
-    va-button::part(button) {
-      height: 4.2rem;
-      width: fit-content;
-      min-width: 78px;
-      margin: 3px;
-    }
-  }
+  width: 100%;
+  // @media (max-width: $small-screen) {
+  word-wrap: break-word;
+  //   }
 
   .unread-column {
     max-width: 15px;
@@ -73,3 +49,27 @@
   }
 }
 
+.thread-list-sort {
+  width: 100%;
+
+  display: flex;
+  align-items: flex-end;
+
+  va-select {
+    margin-right: 16px;
+    min-width: 220px;
+  }
+
+  va-button {
+    display: contents;
+    height: 37px;
+    padding: 0;
+  }
+
+  va-button::part(button) {
+    height: 4.2rem;
+    width: fit-content;
+    min-width: 78px;
+    margin: 3px;
+  }
+}
